closes #1585 changes Integer into Long for memory and disk

+++ coinapi/src/main/java/pl/cyfronet/coin/api/beans/AddAsToWorkflow.java
@@ -26,8 +26,8 @@ public class AddAsToWorkflow {
 	private String asConfigId;
 	private String name;
 	private Float cpu;
-	private Integer memory;
-	private Integer disk;
+	private Long memory;
+	private Long disk;
 
 	public String getAsConfigId() {
 		return asConfigId;
@@ -53,19 +53,19 @@ public class AddAsToWorkflow {
 		this.cpu = cpu;
 	}
 
-	public Integer getMemory() {
+	public Long getMemory() {
 		return memory;
 	}
 
-	public void setMemory(Integer memory) {
+	public void setMemory(Long memory) {
 		this.memory = memory;
 	}
 
-	public Integer getDisk() {
+	public Long getDisk() {
 		return disk;
 	}
 
-	public void setDisk(Integer disk) {
+	public void setDisk(Long disk) {
 		this.disk = disk;
 	}
