+++ coinapi/src/main/java/pl/cyfronet/coin/api/beans/AtomicServiceRequest.java
@@ -34,8 +34,8 @@ public class AtomicServiceRequest {
 	private Boolean published;
 	
 	private Float cpu;
-	private Integer memory;
-	private Integer disk;
+	private Long memory;
+	private Long disk;
 	
 	public String getName() {
 		return name;
@@ -111,19 +111,19 @@ public class AtomicServiceRequest {
 		this.cpu = cpu;
 	}
 
-	public Integer getMemory() {
+	public Long getMemory() {
 		return memory;
 	}
 
-	public void setMemory(Integer memory) {
+	public void setMemory(Long memory) {
 		this.memory = memory;
 	}
 
-	public Integer getDisk() {
+	public Long getDisk() {
 		return disk;
 	}
 
-	public void setDisk(Integer disk) {
+	public void setDisk(Long disk) {
 		this.disk = disk;
 	}
